Refactor: move `private-api` to `macro-support`
