+++ crates/loxide_testsuite/src/lib.rs
@@ -1,9 +1,9 @@
-pub mod __private_api;
 mod macros;
 mod probe;
 mod unittest;
 
 #[doc(hidden)]
 pub mod _macro_support {
+    pub use crate::probe::Probe;
     pub use crate::unittest::tester::source_exec;
 }
