+++ crates/loxide_testsuite/src/macros.rs
@@ -1,14 +1,30 @@
+#[doc(hidden)]
+#[macro_export]
+macro_rules! _function_name {
+    () => {{
+        fn f() {}
+        fn type_name_of_val<T>(_: T) -> &'static str {
+            std::any::type_name::<T>()
+        }
+        let mut name = type_name_of_val(f).strip_suffix("::f").unwrap_or("");
+        while let Some(rest) = name.strip_suffix("::{{closure}}") {
+            name = rest;
+        }
+        name
+    }};
+}
+
 #[macro_export]
 macro_rules! register {
     () => {
-        $crate::__private_api::register($crate::_function_name!());
+        $crate::_macro_support::Probe::register($crate::_function_name!());
     };
 }
 
 #[macro_export]
 macro_rules! probe {
     ($target:expr) => {{
-        $crate::__private_api::probe(::std::format!("{:#?}", $target));
+        $crate::_macro_support::Probe::probe(::std::format!("{:#?}", $target));
         $target
     }};
 }
@@ -16,7 +32,7 @@ macro_rules! probe {
 #[macro_export]
 macro_rules! footprints {
     () => {
-        $crate::__private_api::footprints($crate::_function_name!())
+        $crate::_macro_support::Probe::footprints($crate::_function_name!())
     };
 }
